fix: improve share processing
